requirements.txt
autopep8==1.5.7
beautifulsoup4==4.10.0
certifi==2020.11.8
chardet==3.0.4
idna==2.10
mysqlclient==2.0.1
numpy==1.21.2
pycodestyle==2.7.0
python-dotenv==0.19.0
requests==2.25.0
scipy==1.7.1
selenium==3.141.0
soupsieve==2.0.1
toml==0.10.2
urllib3==1.26.2
